refactor game dialogs
