+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/presentation/gameScreen/ui/components/dialog/RematchPendingDialog.kt
@@ -1,0 +1,39 @@
+package pw.checkers.game.presentation.gameScreen.ui.components.dialog
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+
+@Composable
+internal fun RematchPendingDialog() {
+    Dialog(
+        onDismissRequest = {}
+    ) {
+        Surface(
+            color = MaterialTheme.colorScheme.background,
+            tonalElevation = 8.dp,
+            shape = MaterialTheme.shapes.medium
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                modifier = Modifier.padding(16.dp)
+            ) {
+                Text(
+                    text = "Waiting for response",
+                    style = MaterialTheme.typography.headlineMedium,
+                    modifier = Modifier.padding(top = 8.dp, bottom = 24.dp)
+                )
+
+                CircularProgressIndicator()
+            }
+        }
+    }
+}
