+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/presentation/gameScreen/ui/components/icon/pawns/WhiteQueen.kt
@@ -1,4 +1,4 @@
-package pw.checkers.game.presentation.gameScreen.ui.components.icons.pawns
+package pw.checkers.game.presentation.gameScreen.ui.components.icon.pawns
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.PathFillType.Companion.EvenOdd
@@ -10,7 +10,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.graphics.vector.ImageVector.Builder
 import androidx.compose.ui.graphics.vector.path
 import androidx.compose.ui.unit.dp
-import pw.checkers.game.presentation.gameScreen.ui.components.icons.Pawns
+import pw.checkers.game.presentation.gameScreen.ui.components.icon.Pawns
 
 val Pawns.WhiteQueen: ImageVector
     get() {
